Additionally uses the Auth Code flow
This branch updates the SpotifyCredentials.java for the Authorization Code + PKCE flow (accesses a Spotify account to play songs), which is then used as a dependency for the newly created SpotifyAPI.java class, which handles the Spotify Web API requests. This effectively leaves all the Spotify Web API handling for PlayerLogic.java in the one dependency, SpotifyAPI.java.

I was looking at Dependency Injection this month, which is somewhat present in the code so as to prepare for its integrated it to Spring Boot, but this project has taken too much precedence amidst looking for a Front-end development job. I feel I honestly understand some of the value of DI and its power, updatability, and testing benefits in comparing it with an implementation without it, especially with bigger applications though.

The application at present plays, pauses, and searches for tracks connected to a Spotify account. The last updates I plan to do are 1) updating to the BoxLayout for the Display.java class (I know) 2) to easily add previous and forward track buttons, 3) and getting the 'close' button to work well without affecting the dragging of the JFrame.